Change creation of the connection and services
I was trying too hard to make the creation of the connection easy, but
in fact it just made it impossible to create it the way you might want
to with all of the Faraday settings. This removes the stuff that creates
the connection and relies on the user to inject the middleware into
their Faraday connections manually.

I was also trying to be too smart about service testing, where it could
only compare the service against host and path. In fact, there are
probably other things, like the request method and others. This makes
the service pass a block for evaluating whether or not it applies.