Initial commit (#1)
* getting somewhere

* Moving along with the request implementation

* Add dashboard

* Add logging

This adds logging to a logger passed in with the client. It also changes
the code to use named parameters throughout.

* working, with proper testing

* rename the gem to breakers

* renaming things

* license and readme

* grammar correction

* Change creation of the connection and services

I was trying too hard to make the creation of the connection easy, but
in fact it just made it impossible to create it the way you might want
to with all of the Faraday settings. This removes the stuff that creates
the connection and relies on the user to inject the middleware into
their Faraday connections manually.

I was also trying to be too smart about service testing, where it could
only compare the service against host and path. In fact, there are
probably other things, like the request method and others. This makes
the service pass a block for evaluating whether or not it applies.

* re-raise the error

* make more of the service configurable

* remove cruft

* remove the dash

* utc times and configurable redis prefix

* update README

* improve test coverage

Remove some code that was here just to support the dashboard and
increase the test coverage to 99%.

* bump version to 0.1.0